Split dlc schema
